requirements.txt
numpy
matplotlib
pyyaml
natsort
scipy
opencv-python